+++ bin/actions/0004444-105.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-#: ◤ + -> - Move active window one screen to the left
+#: ◤ + Left - Move active window one screen to the left
 
 xdotool sleep 0.1
 xdotool key Super+Shift+Left