+++ bin/actions/0004444-106.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-#: ◤ + -> - Move active window to screen right of current
+#: ◤ + Right - Move active window to screen right of current
 
 xdotool sleep 0.1
 xdotool key Super+Shift+Right