+++ bin/actions/0220022-28.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+#: ◢ + ENTER - Click left mouse button three times (select line).
+#: Mouse+Key must be "hit": released immediately
+
+sleep 0.5
+xdotool click --repeat 3 1