+++ bin/actions/1010101-103.sh
@@ -13,3 +13,6 @@ echo "Volume++" | aosd_cat -R white -u 300 --y-offset=40 -B black -b 240 \
 --padding=24 -x 5 -n "Ubuntu 20" -p 1
 
 xdotool key Down
+
+sleep 0.1
+xdotool key Escape  # hide context menu popup