+++ bin/actions/1010101-30.sh
@@ -4,3 +4,6 @@
 
 xdotool sleep 0.1
 xdotool key ctrl+a
+
+sleep 0.1
+xdotool key Escape  # hide context menu popup