+++ CHANGELOG.md
@@ -3,9 +3,10 @@ xmchord Changelog
 
 V0.4.1 - Not released yet
 -------------------------
-* Fix action shell scripts style issues (fulfill google convention 100% now)
+* Fix action shell scripts style issues (google convention)
 * Add shellcheck continuous inspection of action shell scripts to travis config   
 * Add "Build tools and Continuous Integration" section into README.md
+* Add action: ◢ + ENTER = Click left mouse button three times (select line)
 
 V0.4.0 - 2021/01/12
 -------------------